Use off_t more consistently for file offsets and file size
- Change the seek and tell methods to off_t in the first place,
  and use fseeko/ftello throughout.
- Change class members holding file sizes and offsets to off_t,
  use matching casts.
- Use strtoll rather than sscanf to read octal file length
  in tar archives.  It's faster and typesafe.
- Use off_t as argument types for progress bar functions.

Additionally:

- Reformat io_stream_memory.h, which was next to unreadable
  due to bad formatting.

+++ UserSettings.cc
@@ -33,7 +33,7 @@ public:
   ssize_t write (const void *buffer, size_t len) {return 0;}
   ssize_t peek (void *buffer, size_t len) {return 0;}
   long tell () {return 0;}
-  int seek (long, io_stream_seek_t) {return 0;}
+  off_t seek (off_t, io_stream_seek_t) {return 0;}
   int error () {return 0;}
   void operator << (std::string);
   void operator << (const char *);
