+++ compress_gz.h
@@ -32,8 +32,8 @@ public:
   virtual ssize_t write (const void *buffer, size_t len);
   /* read data without removing it from the class's internal buffer */
   virtual ssize_t peek (void *buffer, size_t len);
-  virtual long tell ();
-  virtual int seek (long where, io_stream_seek_t whence);
+  virtual off_t tell ();
+  virtual off_t seek (off_t where, io_stream_seek_t whence);
   /* try guessing this one */
   virtual int error ();
   /* Find out the next stream name -
@@ -80,7 +80,7 @@ private:
   char *msg;			/* error message */
   int transparent;		/* 1 if input file is not a .gz file */
   char mode;			/* 'w' or 'r' */
-  long startpos;		/* start of compressed data in file (header skipped) */
+  off_t startpos;		/* start of compressed data in file (header skipped) */
 };
 
 #endif /* SETUP_COMPRESS_GZ_H */
